Refactor views with shared date utilities
