+++ refrigerator_management/Views/ShoppingListView.swift
@@ -15,12 +15,6 @@ struct ShoppingListView: View {
     @State private var deleteOffsets: IndexSet? = nil
     @State private var showingCartConfirm = false
 
-    private static let dateFormatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.dateFormat = "M/d"
-        return formatter
-    }()
-
     var body: some View {
         NavigationView {
             ZStack {
@@ -254,27 +248,10 @@ struct ShoppingListView: View {
     }
 
     private func color(for date: Date) -> Color {
-        let calendar = Calendar.current
-        let today = calendar.startOfDay(for: Date())
-        if date < today {
-            return .red
-        } else if calendar.isDateInTomorrow(date) {
-            return .orange
-        } else {
-            return .gray
-        }
+        DateUtils.color(for: date)
     }
 
     private func dateLabel(for date: Date) -> String {
-        let calendar = Calendar.current
-        if date < calendar.startOfDay(for: Date()) {
-            return "期限切れ"
-        } else if calendar.isDateInToday(date) {
-            return "本日まで"
-        } else if calendar.isDateInTomorrow(date) {
-            return "明日まで"
-        } else {
-            return "期限: \(Self.dateFormatter.string(from: date))"
-        }
+        DateUtils.label(for: date)
     }
 }
